Reorganizacao dos arquivos referentes a watch later e favorite
