+++ src/controller/savedVideosController.py
@@ -1,18 +1,17 @@
 from fastapi import APIRouter, HTTPException, Depends, Query
 from sqlalchemy.orm import Session
-from model import watchLaterModel
-from domain import watchLaterSchema
+from domain import savedVideosSchema
 from database import get_db
-
+from repository import savedVideosRepository
+from starlette.responses import JSONResponse
 
 WatchLater = APIRouter(
  prefix="/watch-later"
 )
 
-
 @WatchLater.post("/")
-def add_to_watch_later(watch_later: watchLaterSchema.WatchLaterCreate, db: Session = Depends(get_db)):
-   return watchLaterModel.create_watch_later(db=db, watch_later=watch_later)
+def add_to_watch_later(watch_later: savedVideosSchema.WatchLaterCreate, db: Session = Depends(get_db)):
+   return savedVideosRepository.create_watch_later(db=db, watch_later=watch_later)
 
 
 @WatchLater.delete("/{video_id}")
@@ -20,13 +19,13 @@ def remove_from_watch_later(video_id: str, user_id: str = Query(...), db: Sessio
    print(f"Attempting to remove video_id={video_id} for user_id={user_id}")
    user_id = user_id.strip()  # Certifique-se de que o `user_id` não contém espaços ou quebras de linha
    video_id = video_id.strip()  # Certifique-se de que o `video_id` não contém espaços ou quebras de linha
-   watchLaterModel.remove_watch_later(db=db, video_id=video_id, user_id=user_id)
+   savedVideosRepository.remove_watch_later(db=db, video_id=video_id, user_id=user_id)
    return {"message": "Removed from watch later list"}
 
 
 @WatchLater.get("/status/{video_id}")
 def check_watch_later(video_id: str, user_id: str = Query(...), db: Session = Depends(get_db)):
    print(f"Checking watch later status for video_id={video_id}, user_id={user_id}")
-   status = watchLaterModel.check_watch_later_status(db=db, video_id=video_id, user_id=user_id)
+   status = savedVideosRepository.check_watch_later_status(db=db, video_id=video_id, user_id=user_id)
    print(f"Status for video_id={video_id}, user_id={user_id} is {status}")
    return {"status": status}
